REFACTOR: unfollow 기능 엔드포인트 수정
- DELETE api/users/{profileId}/follow -> DELETE api/follows/{followId}로 변경
- FollowService에서 followId와 userDetails의 userId 기반으로 삭제 로직 변경

jira ref: #S12P11A209-52

+++ devoot-backend/src/main/java/com/gamee/devoot_backend/follow/service/FollowService.java
@@ -10,6 +10,7 @@ import com.gamee.devoot_backend.common.pageutils.CustomPage;
 import com.gamee.devoot_backend.follow.dto.FollowUserDto;
 import com.gamee.devoot_backend.follow.entity.Follow;
 import com.gamee.devoot_backend.follow.exception.FollowCannotFollowSelfException;
+import com.gamee.devoot_backend.follow.exception.FollowNotAuthorizedException;
 import com.gamee.devoot_backend.follow.exception.FollowRelationshipAlreadyExists;
 import com.gamee.devoot_backend.follow.exception.FollowRelationshipNotFound;
 import com.gamee.devoot_backend.follow.exception.FollowRequestPendingException;
@@ -54,8 +55,8 @@ public class FollowService {
 
 		// 알림 생성
 		Notification notification = Notification.builder()
-			.toUser(followedId)
-			.fromUser(followerId)
+			.toUserId(followedId)
+			.fromUserId(followerId)
 			.followId(savedFollow.getId())
 			.hasRead(false)
 			.build();
@@ -63,18 +64,15 @@ public class FollowService {
 	}
 
 	@Transactional
-	public void deleteFollower(String followerProfileId, String followedProfileId) {
-		User[] users = getUsersByProfileIds(followerProfileId, followedProfileId);
-		User followerUser = users[0];
-		User followedUser = users[1];
-		Long followerId = followerUser.getId();
-		Long followedId = followedUser.getId();
-
+	public void deleteFollower(Long followId, Long currentUserId) {
 		// 이미 팔로우 관계가 존재하는지 확인
-		Follow existingFollow = followRepository.findByFollowerIdAndFollowedId(followerId, followedId)
+		Follow existingFollow = followRepository.findById(followId)
 			.orElseThrow(FollowRelationshipNotFound::new);
-		followRepository.delete(existingFollow);
 
+		if (!existingFollow.getFollowerId().equals(currentUserId)) {
+			throw new FollowNotAuthorizedException();
+		}
+		followRepository.delete(existingFollow);
 		notificationRepository.deleteByFollowId(existingFollow.getId());
 	}
